==> R_script/03_02_model_ARIMA.R <==
# --------------------------------
# Time-series Forecasting - ARIMA
# --------------------------------
library(dplyr)
library(forecast)

# Read data
energy05_23_long <- readRDS("Documents/GitHub/taiwan-renewable-energy-visualisation/R_script/output/energy05_23_long.rds")
ARIMA_forecast2024 <- list()
energy_cols <- c("Hydropower", "Geothermal", "Solar", "Wind", "Biomass", "Waste")

for(e in energy_list){
  ts_hist <- ts(energy05_23_long %>% 
                filter(item==e) %>% 
                pull(value),
                start=c(2005,1), 
                frequency=12)
  fit_arima <- auto.arima(ts_hist) # fit ARIMA
  fc_arima <- forecast(fit_arima, h=12) # forecasting future 12 months
  ARIMA_forecast2024[[e]] <- fc_arima
}

# Store the ARIMA forecast on 2024
saveRDS(ARIMA_forecast2024, "Documents/GitHub/taiwan-renewable-energy-visualisation/R_script/output/ARIMA_forecast2024.rds")
